Alias commands SetGrid<x>Alt allow binding of e.g. KP1-KP0 additionally to 0-1
It is not possible via shotscuts.ini to bind more than one key to
the same command. To let folks use the keys 0-1 along with the keys
KP1-KP0 - as it was possible in v1.4 - needed to create a second set
of command names.

These SetGrid<x>Alt commands do not actually call new functions, they
shader them with SetGrid<x>. IOW the SetGrid<x>Alt commands are command
aliases.